requirements.txt
# Core dependencies
asyncio-throttle>=1.0.2
aiohttp>=3.8.0
requests>=2.28.0
httpx[http2]>=0.24.0
h2>=4.0.0
python-coinmarketcap
cachetools

# Cryptocurrency exchange APIs
ccxt>=4.0.0

# Web scraping and automation
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Cryptography and security
cryptography>=41.0.0
pynacl>=1.5.0

# Data processing and visualization  
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0

# Terminal UI and colors
colorama>=0.4.6
rich>=13.0.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0

# Blockchain and Web3
web3>=6.0.0
solana>=0.30.0
requests-oauthlib>=1.3.0
base58>=2.1.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.0.0 
